Preprocessing, language modeling and classification RNN
GRU-based RNN for language modeling, that can be used as pre-training for a classifier or vice-versa. A number of pre-processing helper functions prepare Reddit comment data for the neural network analysis. Pending: soft-coding the number of output classes for the classifier and creating the possibility of pre-training on a different dataset than the one used in the main task.