R code and results of regression analyses
This folder mainly includes R code, data files and results associated with regression analyses reported in the manuscript. 

Some data files are obsolete and have been included to prevent errors from obsolete sections of the code (the code will be reformatted as soon as possible to resolve the need for doing this). See readme.txt for more details.